Add validation pattern assertion types
